split  PSP and containers
