+++ checkov/terraform/checks/resource/kubernetes/PrivilegedContainer.py
@@ -2,15 +2,15 @@ from checkov.common.models.enums import CheckCategories, CheckResult
 from checkov.terraform.checks.resource.base_resource_check import BaseResourceCheck
 
 
-class PrivilegedContainersPSP(BaseResourceCheck):
+class PrivilegedContainers(BaseResourceCheck):
 
     def __init__(self):
         # CIS-1.3 1.7.1
         # CIS-1.5 5.2.1
         name = "Do not admit privileged containers"
-        id = "CKV_K8S_2"
+        id = "CKV_K8S_16"
 
-        supported_resources = ['kubernetes_pod', 'kubernetes_pod_security_policy']
+        supported_resources = ['kubernetes_pod']
         categories = [CheckCategories.GENERAL_SECURITY]
         super().__init__(name=name, id=id, categories=categories, supported_resources=supported_resources)
 
@@ -24,11 +24,7 @@ class PrivilegedContainersPSP(BaseResourceCheck):
                     if context.get("privileged") == [True]:
                         self.evaluated_keys = [f'spec/[0]/container/[{idx}]/security_context/[0]/privileged']
                         return CheckResult.FAILED
-        # for psp
-        if spec.get("privileged") == [True]:
-            self.evaluated_keys = ['spec/[0]/privileged']
-            return CheckResult.FAILED
         return CheckResult.PASSED
 
 
-check = PrivilegedContainersPSP()
+check = PrivilegedContainers()
