+++ tests/terraform/checks/resource/kubernetes/example_PrivilegedContainersPSP/psp.tf
